Update Optiz for Task 1
