+++ evaluate/web_ui/__init__.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python3
+
+from .monitor import AgentMonitor
+from .server import (
+    start_dashboard_server,
+    log_task_start,
+    log_interaction,
+    log_task_complete,
+    log_vlm_call,
+    start_disambiguation_web,
+    monitor
+)
+
+__all__ = [
+    'AgentMonitor',
+    'start_dashboard_server',
+    'log_task_start',
+    'log_interaction',
+    'log_task_complete',
+    'log_vlm_call',
+    'start_disambiguation_web',
+    'monitor'
+]
