+++ evaluate/web_ui/static/css/dashboard.css
@@ -1,0 +1,363 @@
+* { 
+    margin: 0; 
+    padding: 0; 
+    box-sizing: border-box; 
+}
+
+body {
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+    background: #f5f7fa;
+    color: #333;
+    line-height: 1.6;
+}
+
+.header {
+    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+    color: white;
+    padding: 20px;
+    text-align: center;
+    box-shadow: 0 4px 6px rgba(0,0,0,0.1);
+}
+
+.header-content {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 20px;
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+.logo {
+    height: 60px;
+    width: auto;
+    object-fit: contain;
+    border-radius: 8px;
+    background: rgba(255, 255, 255, 0.1);
+    padding: 5px;
+}
+
+.header-text {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.header h1 {
+    font-size: 2em;
+    margin-bottom: 10px;
+    margin: 0;
+}
+
+.header p {
+    margin: 5px 0 0 0;
+    opacity: 0.9;
+}
+
+.status-bar {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: 20px;
+    padding: 20px;
+    background: white;
+    margin: 0 20px 20px 20px;
+    border-radius: 10px;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+}
+
+.status-card {
+    text-align: center;
+    padding: 15px;
+    border-radius: 8px;
+    background: #f8f9ff;
+    border-left: 4px solid #667eea;
+}
+
+.status-card h3 { 
+    color: #667eea; 
+    font-size: 1.1em; 
+    margin-bottom: 8px; 
+}
+
+.status-card .value { 
+    font-size: 1.5em; 
+    font-weight: bold; 
+    color: #2c3e50; 
+}
+
+.status-card .label { 
+    font-size: 0.9em; 
+    color: #7f8c8d; 
+}
+
+.progress-bar {
+    width: 100%;
+    height: 8px;
+    background: #ecf0f1;
+    border-radius: 4px;
+    overflow: hidden;
+    margin: 10px 0;
+}
+
+.progress-fill {
+    height: 100%;
+    background: linear-gradient(90deg, #667eea, #764ba2);
+    transition: width 0.3s ease;
+}
+
+.main-content {
+    display: grid;
+    grid-template-columns: 1fr 2fr 1fr;
+    gap: 20px;
+    margin: 0 20px;
+}
+
+.panel {
+    background: white;
+    border-radius: 10px;
+    padding: 20px;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+}
+
+.panel h2 {
+    color: #667eea;
+    margin-bottom: 15px;
+    font-size: 1.3em;
+    border-bottom: 2px solid #f1f3f4;
+    padding-bottom: 10px;
+}
+
+.task-history {
+    max-height: 400px;
+    overflow-y: auto;
+}
+
+.task-item {
+    padding: 12px;
+    margin: 8px 0;
+    border-radius: 8px;
+    border-left: 4px solid;
+    background: #f8f9ff;
+}
+
+.task-item.completed { 
+    border-left-color: #27ae60; 
+}
+
+.task-item.failed { 
+    border-left-color: #e74c3c; 
+}
+
+.task-item.running { 
+    border-left-color: #f39c12; 
+    animation: pulse 2s infinite; 
+}
+
+.task-item .task-name { 
+    font-weight: bold; 
+    margin-bottom: 5px; 
+}
+
+.task-item .task-meta { 
+    font-size: 0.85em; 
+    color: #7f8c8d; 
+}
+
+.interaction-log {
+    max-height: 500px;
+    overflow-y: auto;
+}
+
+.interaction-item {
+    padding: 10px;
+    margin: 5px 0;
+    border-radius: 6px;
+    border-left: 3px solid;
+    background: #fafbfc;
+}
+
+.interaction-item.action { 
+    border-left-color: #3498db; 
+}
+
+.interaction-item.vlm_call { 
+    border-left-color: #9b59b6; 
+}
+
+.interaction-item.navigate { 
+    border-left-color: #e67e22; 
+}
+
+.interaction-item.disambiguation_complete { 
+    border-left-color: #9b59b6; 
+}
+
+.disambiguation-history {
+    max-height: 500px;
+    overflow-y: auto;
+}
+
+.disambiguation-item {
+    padding: 15px;
+    margin: 10px 0;
+    border-radius: 8px;
+    background: #f8f9ff;
+    border: 1px solid #e1e5e9;
+    position: relative;
+}
+
+.disambiguation-item.completed {
+    border-left: 4px solid #27ae60;
+}
+
+.disambiguation-item.auto-selected {
+    border-left: 4px solid #f39c12;
+}
+
+.disambiguation-item .disambiguation-header {
+    font-weight: bold;
+    margin-bottom: 8px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.disambiguation-item .disambiguation-time {
+    font-size: 0.8em;
+    color: #7f8c8d;
+}
+
+.disambiguation-item .disambiguation-details {
+    font-size: 0.9em;
+    color: #2c3e50;
+    margin: 5px 0;
+}
+
+.disambiguation-item .selected-choice {
+    background: #e8f5e8;
+    padding: 8px;
+    border-radius: 4px;
+    margin-top: 8px;
+    border-left: 3px solid #27ae60;
+}
+
+.disambiguation-candidates {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    gap: 8px;
+    margin-top: 8px;
+}
+
+.disambiguation-candidate {
+    padding: 6px;
+    background: #f1f3f4;
+    border-radius: 4px;
+    font-size: 0.8em;
+    text-align: center;
+}
+
+.disambiguation-candidate.selected {
+    background: #e8f5e8;
+    font-weight: bold;
+}
+
+.interaction-item .timestamp {
+    float: right;
+    font-size: 0.8em;
+    color: #95a5a6;
+}
+
+.disambiguation-panel {
+    margin-top: 20px;
+    padding: 20px;
+    background: #fff3cd;
+    border: 1px solid #ffeaa7;
+    border-radius: 10px;
+}
+
+.candidate-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 15px;
+    margin: 15px 0;
+}
+
+.candidate-card {
+    border: 2px solid #ddd;
+    border-radius: 8px;
+    padding: 10px;
+    text-align: center;
+    cursor: pointer;
+    transition: all 0.3s ease;
+}
+
+.candidate-card:hover { 
+    transform: translateY(-2px); 
+    box-shadow: 0 4px 8px rgba(0,0,0,0.1); 
+}
+
+.candidate-card.high-confidence { 
+    border-color: #27ae60; 
+}
+
+.candidate-card.medium-confidence { 
+    border-color: #f39c12; 
+}
+
+.candidate-card.low-confidence { 
+    border-color: #e74c3c; 
+}
+
+.candidate-image {
+    width: 100%;
+    height: 150px;
+    object-fit: cover;
+    border-radius: 4px;
+    margin-bottom: 10px;
+}
+
+.select-btn {
+    background: #667eea;
+    color: white;
+    border: none;
+    padding: 8px 16px;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 0.9em;
+}
+
+.select-btn:hover { 
+    background: #5a6fd8; 
+}
+
+@keyframes pulse {
+    0% { opacity: 1; }
+    50% { opacity: 0.7; }
+    100% { opacity: 1; }
+}
+
+.hidden { 
+    display: none; 
+}
+
+@media (max-width: 768px) {
+    .main-content { 
+        grid-template-columns: 1fr; 
+    }
+    .status-bar { 
+        grid-template-columns: repeat(2, 1fr); 
+    }
+    .header-content {
+        flex-direction: column;
+        gap: 15px;
+    }
+    .logo {
+        height: 50px;
+    }
+    .header h1 {
+        font-size: 1.5em;
+    }
+    .disambiguation-candidates {
+        grid-template-columns: 1fr;
+    }
+}
